Revert "[mlir python] Add nanobind support for standalone dialects. (#117922)"
This reverts commit afe75b4d5fcebd6fdd292ca1797de1b35cb687b0.
