+++ mlir/python/requirements.txt
@@ -1,4 +1,3 @@
-nanobind>=2.0, <3.0
 numpy>=1.19.5, <=2.1.2
 pybind11>=2.10.0, <=2.13.6
 PyYAML>=5.4.0, <=6.0.1
