+++ utils/bazel/WORKSPACE
@@ -148,24 +148,6 @@ maybe(
     url = "https://github.com/pybind/pybind11/archive/v2.10.3.zip",
 )
 
-maybe(
-    http_archive,
-    name = "robin_map",
-    strip_prefix = "robin-map-1.3.0",
-    sha256 = "a8424ad3b0affd4c57ed26f0f3d8a29604f0e1f2ef2089f497f614b1c94c7236",
-    build_file = "@llvm-raw//utils/bazel/third_party_build:robin_map.BUILD",
-    url = "https://github.com/Tessil/robin-map/archive/refs/tags/v1.3.0.tar.gz",
-)
-
-maybe(
-    http_archive,
-    name = "nanobind",
-    build_file = "@llvm-raw//utils/bazel/third_party_build:nanobind.BUILD",
-    sha256 = "bfbfc7e5759f1669e4ddb48752b1ddc5647d1430e94614d6f8626df1d508e65a",
-    strip_prefix = "nanobind-2.2.0",
-    url = "https://github.com/wjakob/nanobind/archive/refs/tags/v2.2.0.tar.gz",
-)
-
 load("@rules_python//python:repositories.bzl", "py_repositories", "python_register_toolchains")
 
 py_repositories()
