--- utils/bazel/third_party_build/nanobind.BUILD
@@ -1,25 +1,0 @@
-cc_library(
-    name = "nanobind",
-    srcs = glob(
-        [
-            "src/*.cpp",
-        ],
-        exclude = ["src/nb_combined.cpp"],
-    ),
-    defines = [
-        "NB_BUILD=1",
-        "NB_SHARED=1",
-    ],
-    includes = ["include"],
-    textual_hdrs = glob(
-        [
-            "include/**/*.h",
-            "src/*.h",
-        ],
-    ),
-    visibility = ["//visibility:public"],
-    deps = [
-        "@robin_map",
-        "@rules_python//python/cc:current_py_cc_headers",
-    ],
-)
