--- utils/bazel/third_party_build/robin_map.BUILD
@@ -1,12 +1,0 @@
-cc_library(
-    name = "robin_map",
-    hdrs = [
-        "include/tsl/robin_growth_policy.h",
-        "include/tsl/robin_hash.h",
-        "include/tsl/robin_map.h",
-        "include/tsl/robin_set.h",
-    ],
-    includes = ["."],
-    strip_include_prefix = "include",
-    visibility = ["//visibility:public"],
-)
